fix imports of test functions/strategies
